chore: rename component file #402
