+++ frontend/src/views/projects/roles/TeacherProjectView.vue
@@ -5,7 +5,7 @@ import { type Teacher } from '@/types/users/Teacher.ts';
 import { type Project } from '@/types/Project.ts';
 import ProjectInfo from '@/components/projects/ProjectInfo.vue';
 import ProjectMeter from '@/components/projects/ProjectMeter.vue';
-import DownloadCSV from '@/components/projects/DownloadCSV.vue';
+import DownloadCSVButton from '@/components/projects/DownloadCSVButton.vue';
 
 /* Props */
 defineProps<{
@@ -37,9 +37,9 @@ defineProps<{
         <div class="col-12 md:col-4">
             <ProjectMeter :project="project" />
         </div>
-        <div class="col-12">
+        <div class="col-12 md:col-4">
             <template v-if="project !== null">
-                <DownloadCSV :project="project" />
+                <DownloadCSVButton :project="project" />
             </template>
         </div>
     </div>
